01/django init (#11)
Initializes the project

What does this commit do?
===============

This commit adds  Django as a dependency, sets up the 
directory structure with two initial apps, and configures 
Django's settings with both a development and production 
settings file.

This commit also adds a Makefile, requirements.txt, and 
setup.py

Why are we doing this?
==================

The requirements.txt file uses setup.py to install the package ('-e .'). This
allows us to control dependencies in setup.py, and keep requirements.txt 
for development purposes. 

I will be adding a requirements-dev.txt file in a future commit.
soon.