Add DRF and setup API structure (#12)
What does this commit do?
=========================

This commit adds Django Rest Framework as a dependency and creates some
basic url patterns in the base urls.py file, which references each apps'
urls.py file. There are no views hooked up to the urls yet.

This commit also renames the 'auth' app to 'authentication'.

Why are we doing this?
=====================

The 'auth' app was conflicting with 'django.contrib.auth'.