Adds batch v1 route, regression tests, typing (#13)
What does this commit do?
========================

This commit adds the original solver algorithm and hooks it up to a
route at /v1/ to satisfy backwards-compatibility. It also adds two
simple regression tests to keep it in-line. Lastly, this commit adds
some type safety, including two dev dependencies on stub packages for
django and django-rest-framework.

Closes #3 

Why are we doing this?
======================

Backwards compatibility is a core requirement of the project. One of
Python's weaknesses in enterprise software is a lack of type safety, so
I want to make use of the new typing features in Python 3.