Adds custom user/manager, some simple unit tests (#14)
What does this PR do?
=========================

This commit adds a custom user/manager implementation which is semi-typed,
and some simple unit tests for user-creation. It also removes the Django
admin app, and renames a target in the Makefile

Closes #6 

Why are we doing this?
======================

Due to an in-progress PR in Django (#12405), the django-stubs package
cannot handle generic subtyping of Manager or QuerySet. I got around
this by flagging the problem-lines with "type: ignore".