Enforces authentication by default, and exempts the v1 batch route
What does this commit do?
=========================

Adds DRF's IsAuthenticated permissions to the default permissions,
enforcing it by default across all views. It also exempts the v1/batch
view to maintain backwards compatibility. The CreateAccount view has
already been exempt.

Why are we doing this?
======================

There's no reason for this API to be public, so we can at least control
access in the future by establishing a foundation of strict permissions.